Move admin-handler related render methods to templates/admin.dart
