+++ app/lib/frontend/templates.dart
@@ -656,23 +656,6 @@ class TemplateService {
     );
   }
 
-  /// Renders the `views/uploader_confirmed.mustache` template.
-  String renderUploaderConfirmedPage(String package, String uploaderEmail) {
-    final String content = templateCache.renderTemplate('uploader_confirmed', {
-      'package': package,
-      'uploader_email': uploaderEmail,
-    });
-    return renderLayoutPage(PageType.package, content,
-        title: 'Uploader confirmed', includeSurvey: false);
-  }
-
-  /// Renders the `views/authorized.mustache` template.
-  String renderAuthorizedPage() {
-    final String content = templateCache.renderTemplate('authorized', {});
-    return renderLayoutPage(PageType.package, content,
-        title: 'Pub Authorized Successfully', includeSurvey: false);
-  }
-
   /// Renders the `views/index.mustache` template.
   String renderErrorPage(
       String title, String message, List<PackageView> topPackages) {
